LayoutTests/http/tests/referrer-policy-img/no-referrer-when-downgrade/cross-origin-http-UpgradeMixedContent.https.html

<!DOCTYPE html><!-- webkit-test-runner [ UpgradeMixedContentEnabled=true ] -->
<html>
<head>
<script src="/js-test-resources/js-test.js"></script>
</head>
<body>
<script>
description("Tests the behavior of no-referrer-when-downgrade referrer policy when cross origin from HTTPS to HTTP.");
jsTestIsAsync = true;

if (window.internals) {
    internals.settings.setAllowDisplayOfInsecureContent(true);
    internals.setTrackingPreventionEnabled(false);
}

if (window.testRunner) {
    testRunner.setStatisticsShouldDowngradeReferrer(false, function () {
        // Without HTTPS upgrade this should be the empty string (""), but with HTTPS upgrade it should be the full URL
        let expected = document.location.href;
        let img = new Image();
        img.referrerPolicy = "no-referrer-when-downgrade";
        img.onload = function() { testPassed('referrer is "' + expected + '"'); finishJSTest(); }
        img.onerror = function() { testFailed("referrer is not as expected"); finishJSTest(); }
        img.src = "http://127.0.0.1:8000/referrer-policy/resources/image.py?expected=" + expected;
        if (window.internals)
            internals.setTrackingPreventionEnabled(true);
    });
}

</script>
</body>
</html>
