LayoutTests/http/tests/referrer-policy-img/no-referrer-when-downgrade/cross-origin-http-UpgradeMixedContent.https-expected.txt

Tests the behavior of no-referrer-when-downgrade referrer policy when cross origin from HTTPS to HTTP.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS referrer is "https://127.0.0.1:8443/referrer-policy-img/no-referrer-when-downgrade/cross-origin-http-UpgradeMixedContent.https.html"
PASS successfullyParsed is true

TEST COMPLETE
